整理代码结构
