+++ ZLOnlineAnalysisSystem/pom.xml
@@ -6,7 +6,7 @@
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>2.1.2.RELEASE</version>
-		<relativePath/> <!-- lookup parent from repository -->
+		<relativePath/> <!-- lookup parent from dao -->
 	</parent>
 	<groupId>com.sk</groupId>
 	<artifactId>zl</artifactId>
