+++ ZLOnlineAnalysisSystem/src/main/java/com/sk/zl/config/Swagger2.java
@@ -1,0 +1,41 @@
+package com.sk.zl.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+
+/**
+ * @Description : TODO
+ * @Author : Ellie
+ * @Date : 2019/2/18
+ */
+@Configuration
+public class Swagger2 {
+
+    @Value("${swagger.enable}")
+    private boolean enableSwagger;
+
+    @Bean
+    public Docket createRestApi() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .apiInfo(apiInfo())
+                .enable(enableSwagger)
+                .select()
+                .apis(RequestHandlerSelectors.basePackage("com.sk.zl.controller"))
+                .paths(PathSelectors.any())
+                .build();
+    }
+    private ApiInfo apiInfo() {
+        return new ApiInfoBuilder()
+                .title("springboot利用swagger构建api文档")
+                .description("简单优雅的restfun风格")
+                .version("1.0")
+                .build();
+    }
+}
