+++ ZLOnlineAnalysisSystem/src/main/java/com/sk/zl/controller/HelloController.java
@@ -1,0 +1,17 @@
+package com.sk.zl.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * @Description : TODO
+ * @Author : Ellie
+ * @Date : 2019/2/15
+ */
+@RestController
+public class HelloController {
+    @RequestMapping("/")
+    public String index() {
+        return "Greetings from Spring Boot!";
+    }
+}
