+++ ZLOnlineAnalysisSystem/src/main/java/com/sk/zl/entity/PointInfo.java
@@ -1,0 +1,54 @@
+package com.sk.zl.entity;
+
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
+
+import java.util.Date;
+
+@ApiModel
+public class PointInfo {
+
+    @ApiModelProperty("标签")
+    String tag;
+
+    @ApiModelProperty("类型")
+    String type;
+
+    @ApiModelProperty("状态")
+    int status;
+
+    @ApiModelProperty("插入时间")
+    Date insertTime;
+
+    public String getTag() {
+        return tag;
+    }
+
+    public void setTag(String tag) {
+        this.tag = tag;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public int getStatus() {
+        return status;
+    }
+
+    public void setStatus(int status) {
+        this.status = status;
+    }
+
+    public Date getInsertTime() {
+        return insertTime;
+    }
+
+    public void setInsertTime(Date insertTime) {
+        this.insertTime = insertTime;
+    }
+}
