+++ ZLOnlineAnalysisSystem/src/main/java/com/sk/zl/model/JsonRequestModel.java
@@ -1,0 +1,32 @@
+package com.sk.zl.model;
+
+/**
+ * @Description : TODO
+ * @Author : Ellie
+ * @Date : 2019/2/18
+ */
+public class JsonRequestModel {
+    private String id;
+    private String value;
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public void setValue(String value) {
+        this.value = value;
+    }
+
+//    @Override
+//    public String toString() {
+//        return "{id:" + id + ", value:" + value + "}";
+//    }
+}
