--- ZLOnlineAnalysisSystem/src/main/java/com/sk/zl/mybatis/entity/User.java
@@ -1,61 +1,0 @@
-package com.sk.zl.mybatis.entity;
-
-import java.io.Serializable;
-
-public class User implements Serializable {
-    private Integer userId;
-
-    private String userName;
-
-    private String password;
-
-    private String phone;
-
-    private static final long serialVersionUID = 1L;
-
-    public Integer getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Integer userId) {
-        this.userId = userId;
-    }
-
-    public String getUserName() {
-        return userName;
-    }
-
-    public void setUserName(String userName) {
-        this.userName = userName == null ? null : userName.trim();
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password == null ? null : password.trim();
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone == null ? null : phone.trim();
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append(getClass().getSimpleName());
-        sb.append(" [");
-        sb.append("Hash = ").append(hashCode());
-        sb.append(", userId=").append(userId);
-        sb.append(", userName=").append(userName);
-        sb.append(", password=").append(password);
-        sb.append(", phone=").append(phone);
-        sb.append("]");
-        return sb.toString();
-    }
-}
