--- ZLOnlineAnalysisSystem/src/main/java/com/sk/zl/mybatis/mapper/UserMapper.java
@@ -1,83 +1,0 @@
-package com.sk.zl.mybatis.mapper;
-
-import com.sk.zl.mybatis.entity.User;
-import com.sk.zl.mybatis.entity.UserExample;
-import java.util.List;
-import org.apache.ibatis.annotations.Delete;
-import org.apache.ibatis.annotations.DeleteProvider;
-import org.apache.ibatis.annotations.Insert;
-import org.apache.ibatis.annotations.InsertProvider;
-import org.apache.ibatis.annotations.Param;
-import org.apache.ibatis.annotations.Result;
-import org.apache.ibatis.annotations.Results;
-import org.apache.ibatis.annotations.Select;
-import org.apache.ibatis.annotations.SelectProvider;
-import org.apache.ibatis.annotations.Update;
-import org.apache.ibatis.annotations.UpdateProvider;
-import org.apache.ibatis.type.JdbcType;
-
-public interface UserMapper {
-    @SelectProvider(type=UserSqlProvider.class, method="countByExample")
-    long countByExample(UserExample example);
-
-    @DeleteProvider(type=UserSqlProvider.class, method="deleteByExample")
-    int deleteByExample(UserExample example);
-
-    @Delete({
-        "delete from user",
-        "where user_id = #{userId,jdbcType=INTEGER}"
-    })
-    int deleteByPrimaryKey(Integer userId);
-
-    @Insert({
-        "insert into user (user_id, user_name, ",
-        "password, phone)",
-        "values (#{userId,jdbcType=INTEGER}, #{userName,jdbcType=VARCHAR}, ",
-        "#{password,jdbcType=VARCHAR}, #{phone,jdbcType=VARCHAR})"
-    })
-    int insert(User record);
-
-    @InsertProvider(type=UserSqlProvider.class, method="insertSelective")
-    int insertSelective(User record);
-
-    @SelectProvider(type=UserSqlProvider.class, method="selectByExample")
-    @Results({
-        @Result(column="user_id", property="userId", jdbcType=JdbcType.INTEGER, id=true),
-        @Result(column="user_name", property="userName", jdbcType=JdbcType.VARCHAR),
-        @Result(column="password", property="password", jdbcType=JdbcType.VARCHAR),
-        @Result(column="phone", property="phone", jdbcType=JdbcType.VARCHAR)
-    })
-    List<User> selectByExample(UserExample example);
-
-    @Select({
-        "select",
-        "user_id, user_name, password, phone",
-        "from user",
-        "where user_id = #{userId,jdbcType=INTEGER}"
-    })
-    @Results({
-        @Result(column="user_id", property="userId", jdbcType=JdbcType.INTEGER, id=true),
-        @Result(column="user_name", property="userName", jdbcType=JdbcType.VARCHAR),
-        @Result(column="password", property="password", jdbcType=JdbcType.VARCHAR),
-        @Result(column="phone", property="phone", jdbcType=JdbcType.VARCHAR)
-    })
-    User selectByPrimaryKey(Integer userId);
-
-    @UpdateProvider(type=UserSqlProvider.class, method="updateByExampleSelective")
-    int updateByExampleSelective(@Param("record") User record, @Param("example") UserExample example);
-
-    @UpdateProvider(type=UserSqlProvider.class, method="updateByExample")
-    int updateByExample(@Param("record") User record, @Param("example") UserExample example);
-
-    @UpdateProvider(type=UserSqlProvider.class, method="updateByPrimaryKeySelective")
-    int updateByPrimaryKeySelective(User record);
-
-    @Update({
-        "update user",
-        "set user_name = #{userName,jdbcType=VARCHAR},",
-          "password = #{password,jdbcType=VARCHAR},",
-          "phone = #{phone,jdbcType=VARCHAR}",
-        "where user_id = #{userId,jdbcType=INTEGER}"
-    })
-    int updateByPrimaryKey(User record);
-}
