+++ ZLOnlineAnalysisSystem/src/main/java/com/sk/zl/service/PointInfoService.java
@@ -1,0 +1,15 @@
+package com.sk.zl.service;
+
+import com.sk.zl.entity.PointInfo;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public interface PointInfoService {
+
+    List<PointInfo> getPointCurrentStatistic(String type);
+
+    void addPointInfo(PointInfo pointInfo);
+
+}
