--- ZLOnlineAnalysisSystem/src/main/java/com/sk/zl/skdb/entity/HttpPoolProperties.java
@@ -1,68 +1,0 @@
-package com.sk.zl.skdb.entity;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-/**
- * @Description : 连接池配置信息
- * @Author : Ellie
- * @Date : 2019/2/15
- */
-@Component
-@ConfigurationProperties(prefix = "skhttppool")
-public class HttpPoolProperties {
-    private Integer maxTotal;
-    private Integer defaultMaxPerRoute;
-    private Integer connectTimeout;
-    private Integer connectionRequestTimeout;
-    private Integer socketTimeout;
-    private Integer validateAfterInactivity;
-
-    public Integer getMaxTotal() {
-        return maxTotal;
-    }
-
-    public void setMaxTotal(Integer maxTotal) {
-        this.maxTotal = maxTotal;
-    }
-
-    public Integer getDefaultMaxPerRoute() {
-        return defaultMaxPerRoute;
-    }
-
-    public void setDefaultMaxPerRoute(Integer defaultMaxPerRoute) {
-        this.defaultMaxPerRoute = defaultMaxPerRoute;
-    }
-
-    public Integer getConnectTimeout() {
-        return connectTimeout;
-    }
-
-    public void setConnectTimeout(Integer connectTimeout) {
-        this.connectTimeout = connectTimeout;
-    }
-
-    public Integer getConnectionRequestTimeout() {
-        return connectionRequestTimeout;
-    }
-
-    public void setConnectionRequestTimeout(Integer connectionRequestTimeout) {
-        this.connectionRequestTimeout = connectionRequestTimeout;
-    }
-
-    public Integer getSocketTimeout() {
-        return socketTimeout;
-    }
-
-    public void setSocketTimeout(Integer socketTimeout) {
-        this.socketTimeout = socketTimeout;
-    }
-
-    public Integer getValidateAfterInactivity() {
-        return validateAfterInactivity;
-    }
-
-    public void setValidateAfterInactivity(Integer validateAfterInactivity) {
-        this.validateAfterInactivity = validateAfterInactivity;
-    }
-}
