--- ZLOnlineAnalysisSystem/src/main/java/com/sk/zl/web/entity/JsonRequestEntity.java
@@ -1,32 +1,0 @@
-package com.sk.zl.web.entity;
-
-/**
- * @Description : TODO
- * @Author : Ellie
- * @Date : 2019/2/18
- */
-public class JsonRequestEntity {
-    private String id;
-    private String value;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
-
-//    @Override
-//    public String toString() {
-//        return "{id:" + id + ", value:" + value + "}";
-//    }
-}
